Renamed chassis driving commands to more accurately reflect what they do
