+++ commands/chassis/ChassisSetDistance.java
@@ -2,99 +2,121 @@ package org.usfirst.frc4904.standard.commands.chassis;
 
 
 import org.usfirst.frc4904.standard.LogKitten;
+import org.usfirst.frc4904.standard.custom.ChassisController;
+import org.usfirst.frc4904.standard.custom.motioncontrollers.MotionController;
 import org.usfirst.frc4904.standard.custom.sensors.CustomEncoder;
 import org.usfirst.frc4904.standard.custom.sensors.InvalidSensorException;
 import org.usfirst.frc4904.standard.subsystems.chassis.Chassis;
+import edu.wpi.first.wpilibj.command.Command;
 
-public class ChassisSetDistance extends ChassisConstant {
-	protected CustomEncoder[] encoders;
-	protected final ChassisConstant fallbackCommand;
-	protected double distance;
-	protected double[] initialDistances;
+public class ChassisSetDistance extends Command implements ChassisController {
+	protected final ChassisMove chassisMove;
+	protected final MotionController motionController;
+	protected final Command fallbackCommand;
+	protected final double distance;
+	protected final CustomEncoder[] encoders;
 
 	/**
 	 * Constructor.
 	 * This command moves the chassis forward a known distance via a set of encoders.
 	 * The distance is calculated as the average of the provided encoders.
-	 * Control is purely bang (as opposed to bang-bang/pid). The chassis will move at the speed until it passes the point, then will stop.
+	 * The speed is decided by the provided motionController.
 	 *
 	 * @param chassis
 	 * @param distance
 	 *        distance to move in encoder ticks
-	 * @param speed
-	 *        the speed to move at
+	 * @param motionController
 	 * @param fallbackCommand
 	 *        If the sensor fails for some reason, this command will be cancelled, then the fallbackCommand will start
 	 * @param encoders
 	 */
-	public ChassisSetDistance(Chassis chassis, double distance, double speed, ChassisConstant fallbackCommand, CustomEncoder... encoders) {
-		super(chassis, 0.0, speed, 0.0, Double.MAX_VALUE);
+	public ChassisSetDistance(Chassis chassis, double distance, MotionController motionController, Command fallbackCommand, CustomEncoder... encoders) {
+		chassisMove = new ChassisMove(chassis, this, false);
+		this.motionController = motionController;
 		this.encoders = encoders;
 		this.distance = distance;
 		this.fallbackCommand = fallbackCommand;
-		initialDistances = new double[encoders.length];
 	}
-
+	
 	/**
 	 * Constructor.
 	 * This command moves the chassis forward a known distance via a set of encoders.
 	 * The distance is calculated as the average of the provided encoders.
-	 * Control is purely bang (as opposed to bang-bang/pid). The chassis will move at the speed until it passes the point, then will stop.
+	 * The speed is decided by the provided motionController.
 	 *
 	 * @param chassis
 	 * @param distance
 	 *        distance to move in encoder ticks
-	 * @param speed
-	 *        the speed to move at
+	 * @param motionController
 	 * @param encoders
 	 */
-	public ChassisSetDistance(Chassis chassis, double distance, double speed, CustomEncoder... encoders) {
-		this(chassis, distance, speed, null, encoders);
+	public ChassisSetDistance(Chassis chassis, double distance, MotionController motionController, CustomEncoder... encoders) {
+		this(chassis, distance, motionController, null, encoders);
 	}
 
 	@Override
-	protected void initialize() {
-		super.initialize();
-		for (int i = 0; i < encoders.length; i++) {
-			try {
-				initialDistances[i] = encoders[i].getDistance();
-			}
-			catch (InvalidSensorException e) {
-				cancel();
-				return;
-			}
+	public void initialize() {
+		chassisMove.start();
+		try {
+			motionController.reset();
 		}
+		catch (InvalidSensorException e) {
+			chassisMove.cancel();
+			cancel();
+			if (fallbackCommand != null) {
+				fallbackCommand.start();
+			}
+			return;
+		}
+		for (CustomEncoder encoder : encoders) {
+			encoder.reset();
+		}
+		motionController.setSetpoint(distance);
 	}
 
 	@Override
-	protected boolean isFinished() {
-		double distanceSum = 0;
-		for (int i = 0; i < encoders.length; i++) {
-			try {
-				distanceSum += encoders[i].getDistance() - initialDistances[i];
-				LogKitten.d("Encoder " + i + " reads " + encoders[i].getDistance() + " out of " + distance);
-			}
-			catch (InvalidSensorException e) {
-				cancel();
-				if (fallbackCommand != null) {
-					fallbackCommand.start();
-				}
-				return true;
-			}
+	public double getX() {
+		return 0;
+	}
+
+	@Override
+	public double getY() {
+		double speed;
+		try {
+			speed = motionController.get();
 		}
-		double distanceAvg = distanceSum / encoders.length;
-		return distanceAvg >= distance;
+		catch (InvalidSensorException e) {
+			chassisMove.cancel();
+			cancel();
+			if (fallbackCommand != null) {
+				fallbackCommand.start();
+			}
+			speed = 0;
+		}
+		LogKitten.d("MotionProfileSpeed: " + speed);
+		return speed;
+	}
+
+	@Override
+	public double getTurnSpeed() {
+		return 0;
 	}
 
 	@Override
 	protected void end() {
-		super.end();
-		LogKitten.v("Finished traveling " + distance + " units (as set by setDistancePerPulse)");
+		chassisMove.cancel();
 	}
 
 	@Override
+	protected void execute() {}
+
+	@Override
 	protected void interrupted() {
-		super.interrupted();
-		LogKitten.w("Interrupted! " + getName() + " is in undefined location.");
+		end();
+	}
+
+	@Override
+	protected boolean isFinished() {
+		return motionController.onTarget() || !chassisMove.isRunning();
 	}
 }
